Changing number generator to unsigned long long.

+++ sorting/gennums.cpp
@@ -9,16 +9,16 @@ using namespace std;
 int main(int argc, const char ** argv)
 {
   default_random_engine dre;
-  unsigned int num = 0;
-  unsigned int low = 0;
-  unsigned int high = 0;
+  unsigned long long num = 0;
+  unsigned long long low = 0;
+  unsigned long long high = 0;
   istringstream(argv[1]) >> num;
   istringstream(argv[2]) >> low;
   istringstream(argv[3]) >> high;
 
-  uniform_int_distribution<int> distribution(low, high);
+  uniform_int_distribution<unsigned long long> distribution(low, high);
   
-  for(unsigned int i = 0; i < num; ++i)
+  for(unsigned long long i = 0; i < num; ++i)
   {
     cout << distribution(dre) << " ";
   }
